constexpr MouseEvent
