+++ src/events/CMakeLists.txt
@@ -1,4 +1,3 @@
 target_sources(ufpslib PUBLIC
-	mouse_event.cpp
 	mouse_button_event.cpp
 )
