+++ src/events/mouse_event.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <format>
 #include <string>
 
 namespace ufps
@@ -8,16 +9,37 @@ namespace ufps
 class MouseEvent
 {
   public:
-    MouseEvent(float delta_x, float delta_y);
+    constexpr MouseEvent(float delta_x, float delta_y);
 
-    auto delta_x() const -> float;
-    auto delta_y() const -> float;
+    constexpr auto delta_x() const -> float;
+    constexpr auto delta_y() const -> float;
 
-    auto to_string() const -> std::string;
+    constexpr auto to_string() const -> std::string;
 
   private:
     float delta_x_;
     float delta_y_;
 };
+
+constexpr MouseEvent::MouseEvent(float delta_x, float delta_y)
+    : delta_x_(delta_x)
+    , delta_y_(delta_y)
+{
+}
+
+constexpr auto MouseEvent::delta_x() const -> float
+{
+    return delta_x_;
+}
+
+constexpr auto MouseEvent::delta_y() const -> float
+{
+    return delta_y_;
+}
+
+constexpr auto MouseEvent::to_string() const -> std::string
+{
+    return std::format("MouseEvent {} {}", delta_x(), delta_y());
+}
 
 }
